+++ releasenotes.md
@@ -19,6 +19,7 @@
   * The fade to black when quitting *DOOM Retro* from the menu is no longer affected by the player having either a berserk or radiation shielding suit power-up.
   * Minor improvements have been made to the edges of the menu’s background.
   * The skull cursor is no longer positioned incorrectly in the main menu in some rare instances.
+  * The `episode`, `expansion`, `savegame` and `skilllevel` CVARs are now updated as necessary when navigating using the mouse and the `m_poninter` CVAR is `on`.
 * The effects of changing the `r_gamma` and `r_saturation` CVARs are now applied to *DOOM Retro’s* splash screen.
 * The `r_lowpixelsize` CVAR can now also be set to `1×2` or `2×1`.
 * A bug is fixed whereby sometimes the player wouldn’t pick up an item in a liquid sector when the `r_liquid_clipsprites` CVAR was `on`.